requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
cryptography>=40.0.1
pyzmq>=25.0.2
requests>=2.28.2
colorlog>=6.7.0